requirements.txt
pandas
requests
pillow
opencv-python
ImageHash
streamlit
python-dotenv
cloudinary
pymongo
tqdm
selenium
